# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hohaya API",
        "description": "API documentation for Hohaya platform",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/api/properties/{id}": {
            "get": {
                "tags": [
                    "Properties"
                ],
                "summary": "Détail d'une annonce (public)",
                "operationId": "7cea0ed0f808489cbcc61f528935cd41",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'annonce",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail de l'annonce"
                    },
                    "404": {
                        "description": "Annonce introuvable"
                    }
                }
            },
            "put": {
                "tags": [
                    "Properties"
                ],
                "summary": "Modifier les infos d'une annonce (sans médias)",
                "operationId": "19a6c5948a5a419de858916d8a4ab7c6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'annonce",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Champs à modifier (tous optionnels)",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Appartement rénové à Cadjehoun"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Description mise à jour..."
                                    },
                                    "indications_acces": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "quartier": {
                                        "type": "string",
                                        "example": "Cadjehoun"
                                    },
                                    "commune": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "ville": {
                                        "type": "string",
                                        "example": "Cotonou"
                                    },
                                    "pays": {
                                        "type": "string",
                                        "example": "Bénin"
                                    },
                                    "prix_loyer": {
                                        "type": "number",
                                        "example": 80000
                                    },
                                    "devise": {
                                        "type": "string",
                                        "example": "XOF"
                                    },
                                    "type_logement": {
                                        "type": "string",
                                        "enum": [
                                            "appartement",
                                            "maison",
                                            "studio",
                                            "chambre",
                                            "villa",
                                            "duplex"
                                        ]
                                    },
                                    "condition": {
                                        "type": "string",
                                        "example": "Très bon état"
                                    },
                                    "nb_avance": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "caution_electricite": {
                                        "type": "number",
                                        "nullable": true
                                    },
                                    "caution_eau": {
                                        "type": "number",
                                        "nullable": true
                                    },
                                    "nb_pieces": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "date_debut_louer": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "eau_courante": {
                                        "type": "boolean"
                                    },
                                    "electricite": {
                                        "type": "boolean"
                                    },
                                    "gardien": {
                                        "type": "boolean"
                                    },
                                    "parking": {
                                        "type": "boolean"
                                    },
                                    "meuble": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Annonce mise à jour"
                    },
                    "403": {
                        "description": "Non autorisé"
                    },
                    "404": {
                        "description": "Annonce introuvable"
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Properties"
                ],
                "summary": "Supprimer une annonce et tous ses médias (soft delete)",
                "operationId": "acabc512dd07a218f8c1a4cf46a922a4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'annonce",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Annonce supprimée"
                    },
                    "403": {
                        "description": "Non autorisé"
                    },
                    "404": {
                        "description": "Annonce introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/properties/my": {
            "get": {
                "tags": [
                    "Properties"
                ],
                "summary": "Mes annonces (owner uniquement)",
                "operationId": "4c9a3ef88001139ccd7a1954f0232a8c",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mes annonces"
                    },
                    "403": {
                        "description": "Réservé aux propriétaires"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/properties/{id}/status": {
            "patch": {
                "tags": [
                    "Properties"
                ],
                "summary": "Changer le statut d'une annonce (owner)",
                "operationId": "7e215f4cedb3437fed4529c564613acd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "description": "Nouveau statut de l'annonce",
                                        "type": "string",
                                        "example": "loué",
                                        "enum": [
                                            "disponible",
                                            "loué",
                                            "suspendu"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Statut mis à jour"
                    },
                    "403": {
                        "description": "Non autorisé"
                    },
                    "422": {
                        "description": "Statut invalide"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/properties/{id}/verify": {
            "patch": {
                "tags": [
                    "Properties"
                ],
                "summary": "Vérifier ou suspendre une annonce (admin uniquement)",
                "operationId": "27977339e92d5eacceeb28bf2ae555f2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "is_verified"
                                ],
                                "properties": {
                                    "is_verified": {
                                        "description": "true = publier, false = suspendre",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Annonce vérifiée ou suspendue"
                    },
                    "403": {
                        "description": "Réservé aux administrateurs"
                    },
                    "404": {
                        "description": "Annonce introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}